requirements.txt
openai
python-dotenv
matplotlib
seaborn
pandas
numpy